+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - The card_size attribute was added to the tracking_data node
 
+### Reverted
+
+- Congrats activities update of 1.50.0 version
+
 ## [1.50.0]
 
 ### Updated